add confirm delete

+++ src/apiServices.js
@@ -1,6 +1,5 @@
 const baseURL = 'http://localhost:3001';
 
-//============= post questions
 export const postQuestion = async (question) => {
   return fetch(`${baseURL}/questions`, {
     method: 'POST',
@@ -11,21 +10,20 @@ export const postQuestion = async (question) => {
     .catch((error) => console.log(error));
 };
 
-//============= get questions
 export const getQuestions = async () => {
   return fetch(`${baseURL}/questions`)
     .then((result) => result.json())
     .catch((error) => console.log(error));
 };
 
-// TODO -- "delete / edit questions"
-//============= delete questions
-export const deleteQuestion = async (question) => {
-  return fetch(`${baseURL}/questions`, {
-    method: 'DELETE',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(question),
-  })
-    .then((result) => result.json())
-    .catch((error) => console.log(error));
+export const deleteQuestion = async (id) => {
+  if (confirm('Are you sure you want to delete this?') === true) {
+    return fetch(`${baseURL}/questions/${id}`, {
+      method: 'DELETE',
+    })
+      .then((result) => result.json())
+      .catch((error) => console.log(error));
+  } else {
+    return true;
+  }
 };
